refactor to make it easy to build multiple rotations
