+++ new-era-commands/rotations/triage.js
@@ -1,0 +1,12 @@
+const { rotationBuilder } = require("../../services/RotationBuilder");
+
+const triage = rotationBuilder("triage", "maintainerTriageRotationList");
+
+console.log(triage)
+
+const { data, execute } = triage
+
+module.exports = {
+  data,
+  execute,
+};
